requirements.txt
inference-sdk>=0.9.0
streamlit>=1.35.0
reportlab>=4.2.5
sqlalchemy>=2.0.0
pillow>=10.0.0
opencv-python-headless>=4.9.0
plotly>=5.0.0
